# perf: avoid creating unused workers