## Summary

Avoid creating unused workers when the number of tests is less than the default thread count.

This significantly improves performance when running a single test file.

before:
```bash
npx rstest run  1.29s user 0.45s system 85% cpu 2.031 total
```

after:
```bash
npx rstest run  0.72s user 0.26s system 86% cpu 1.137 total
```

## Related Links

<!--- Provide links of related issues or pages -->

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).